# openapi.yaml
openapi: 3.0.3
info:
  title: tt-data-league-api
  version: v1
  description: |
    OpenAPI 3 specification inferred from the `tt-data-league-api-rest` module.
    This file was generated from controller and DTO records and includes
    example payloads and required properties where applicable.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server (inferred)
...
paths:
  /club/find_by_id:
    get:
      tags: [Club API]
      summary: Find club by id
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Club UUID
      responses:
        '200':
          description: Club found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /club/find_by_name:
    get:
      tags: [Club API]
      summary: Find club by exact name
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Club found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /club/find_by_similar_name:
    get:
      tags: [Club API]
      summary: Find clubs by similar name
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of clubs (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClubDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /club/{id}:
    delete:
      tags: [Club API]
      summary: Delete a club by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /club_member/find_by_club_id/{clubId}:
    get:
      tags: [ClubMember API]
      summary: Get club members by club id
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of club members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClubMemberDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /club_member/enriched/find_by_club_id/{clubId}:
    get:
      tags: [ClubMember API]
      summary: Get enriched club members by club id
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of enriched club members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnrichedClubMemberDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /club_member/find_by_practicioner_id/{practicionerId}:
    get:
      tags: [ClubMember API]
      summary: Get club members by practicioner id
      parameters:
        - name: practicionerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of club members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClubMemberDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /match/{id}:
    get:
      tags: [Match API]
      summary: Get match by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /match/enriched/find_matches:
    post:
      tags: [Match API]
      summary: Find enriched matches by season/competition/day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindMatchesRequestBodyDto'
            example:
              season: "2023-2024"
              competitionType: "league"
              competitionCategory: "senior"
              competitionScope: "regional"
              competitionScopeTag: "A"
              competitionGroup: "group-1"
              competitionGender: "male"
              matchDayNumber: 5
      responses:
        '200':
          description: List of enriched matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnrichedMatchDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ClubDto:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000001"
        name:
          type: string
          example: "CTT Example"
        yearRanges:
          type: array
          items:
            type: string
          example: ["2023-2024"]
    ClubMemberDto:
      type: object
      required: [clubId, practicionerId]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000020"
        clubId:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000002"
        practicionerId:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000010"
        yearRanges:
          type: array
          items:
            type: string
          example: ["2023-2024"]
    EnrichedClubMemberDto:
      type: object
      required: [id, club, practicioner]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000030"
        club:
          $ref: '#/components/schemas/ClubDto'
        practicioner:
          $ref: '#/components/schemas/PracticionerDto'
        yearRanges:
          type: array
          items:
            type: string
          example: ["2023-2024"]
    PracticionerDto:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000010"
        firstName:
          type: string
          example: "Juan"
        secondName:
          type: string
          example: "Perez"
        fullName:
          type: string
          example: "Juan Perez"
        birthDate:
          type: string
          format: date
          example: "1990-05-01"
    SeasonPlayerDto:
      type: object
      required: [clubMemberId, licenseId, licenseTag, yearRange]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000050"
        clubMemberId:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000002"
        licenseId:
          type: string
          example: "L-1234"
        licenseTag:
          type: string
          example: "senior"
        yearRange:
          type: string
          example: "2023-2024"
...
    MatchDto:
      type: object
      required: [id, homeTeam, awayTeam, matchDate]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000060"
        homeTeam:
          type: string
          example: "Team A"
        awayTeam:
          type: string
          example: "Team B"
        homeScore:
          type: integer
          example: 3
        awayScore:
          type: integer
          example: 2
        matchDate:
          type: string
          format: date
          example: "2024-06-15"
    CompetitionInfoDto:
      type: object
      properties:
        type:
          type: string
        category:
          type: string
        scope:
          type: string
        scopeTag:
          type: string
        group:
          type: string
        gender:
          type: string
    FindMatchesRequestBodyDto:
      type: object
      required: [season, matchDayNumber]
      properties:
        season:
          type: string
          example: "2023-2024"
        competitionType:
          type: string
          example: "league"
        competitionCategory:
          type: string
          example: "senior"
        competitionScope:
          type: string
          example: "regional"
        competitionScopeTag:
          type: string
          example: "A"
        competitionGroup:
          type: string
          example: "group-1"
        competitionGender:
          type: string
          example: "male"
        matchDayNumber:
          type: integer
          example: 5
    MatchSeasonPlayerResultDto:
      type: object
      properties:
        seasonPlayer:
          $ref: '#/components/schemas/SeasonPlayerDto'
        matchDay:
          type: string
        matchDayNumber:
          type: integer
        matchGamePoints:
          type: string
        matchGamesWon:
          type: integer
        matchPlayerLetter:
          type: string
    EnrichedMatchDto:
      type: object
      required: [id, season, competitionInfo, matchDayNumber]
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-0000-0000-000000000070"
        playerAbcResultDto:
          $ref: '#/components/schemas/MatchSeasonPlayerResultDto'
        playerXyzResultDto:
          $ref: '#/components/schemas/MatchSeasonPlayerResultDto'
        season:
          type: string
          example: "2023-2024"
        competitionInfo:
          $ref: '#/components/schemas/CompetitionInfoDto'
        matchDayNumber:
          type: integer
          example: 5
        uniqueRowMatchId:
          type: string
          example: "row-12345"
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "INTERNAL_ERROR"
        message:
          type: string
          example: "An unexpected server error occurred"
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []